Pagination completed

+++ ReadIt.Angular/src/app/profile/blogs/user-blogs/user-blogs.component.ts
@@ -1,9 +1,11 @@
 import { Component, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { UserBlogService } from 'src/app/core/apiservices/user-blog.service';
 import { Blog } from 'src/app/core/models/blog';
 import { UserAuthService } from 'src/app/core/services/user-auth.service';
+import { ConfirmationModalComponent } from 'src/app/core/shared/confirmation-modal/confirmation-modal.component';
 
 @Component({
   selector: 'app-user-blogs',
@@ -14,7 +16,7 @@ export class UserBlogsComponent {
   userId: number = this.userAuthService.getUserId();
   allBlogs: Blog[] = [];
 
-  displayedColumns: string[] = ['position', 'Title', 'Description', 'Category','Action'];
+  displayedColumns: string[] = ['position', 'Title', 'Category', 'Action'];
   dataSource = new MatTableDataSource<Blog>(this.allBlogs);
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
@@ -23,24 +25,35 @@ export class UserBlogsComponent {
     this.dataSource.paginator = this.paginator;
   }
 
-constructor(private userBlogService: UserBlogService, private userAuthService: UserAuthService){}
-ngOnInit(){
-  this.userBlogService.getAllByUserId(this.userId).subscribe({
-    next : (response) => {
-      this.allBlogs = response.items;
-      this.dataSource.data = response.items;
-    }
-  })
-}
-deleteBlog(id: number) {
-  this.userBlogService.delete(id).subscribe({
-    next: (response) => {
-      this.userBlogService.getAllByUserId(this.userId).subscribe({
-        next : (response) => {
-          this.dataSource.data = response.items;
-        }
-      })
-    }
-  })
-}
+  constructor(private userBlogService: UserBlogService, private userAuthService: UserAuthService, private dialog: MatDialog) { }
+  ngOnInit() {
+    this.userBlogService.getAllByUserId(this.userId).subscribe({
+      next: (response) => {
+        this.allBlogs = response.items;
+        this.dataSource.data = response.items;
+      }
+    })
+  }
+  deleteBlog(id: number) {
+    const dialogRef = this.dialog.open(ConfirmationModalComponent, {
+      width: '400px',
+      data: {
+        title: 'Confirmation',
+        message: 'Are you sure you want to delete?',
+      }
+    })
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.userBlogService.delete(id).subscribe({
+          next: (response) => {
+            this.userBlogService.getAllByUserId(this.userId).subscribe({
+              next: (response) => {
+                this.dataSource.data = response.items;
+              }
+            })
+          }
+        })
+      }
+    })
+  }
 }
